pyproject.toml
[build-system]
requires = [ "setuptools>=64", "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "goobits-matilda-brain"
version = "1.0.3"
description = "A powerful AI assistant for your terminal"
readme = "README.md"
requires-python = ">=3.8"
keywords = [ "ai", "assistant", "terminal", "llm", "chat", "gpt", "claude",]
classifiers = [ "Development Status :: 4 - Beta", "Intended Audience :: Developers", "License :: OSI Approved :: MIT License", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Programming Language :: Python :: 3.13", "Topic :: Software Development :: Libraries :: Python Modules", "Topic :: Communications :: Chat", "Topic :: Scientific/Engineering :: Artificial Intelligence",]
dependencies = [ "click>=8.0.0", "rich-click>=1.7.0", "rich>=13.0.0", "pydantic>=2.0.0", "httpx>=0.24.0", "litellm>=1.0.0", "pyyaml>=6.0.0", "bleach>=6.0.0", "validators>=0.20.0",]
[[project.authors]]
name = "TTT CLI Contributors"

[project.license]
text = "MIT"

[project.optional-dependencies]
local = [ "httpx>=0.24.0",]
dev = [ "pytest>=7.0", "pytest-asyncio", "pytest-cov", "pytest-timeout", "pytest-xdist>=3.0.0", "pytest-sugar>=0.9.0", "pytest-clarity>=1.0.0", "pytest-html>=3.0.0", "pytest-json-report>=1.5.0", "black", "ruff", "mypy", "types-PyYAML",]
publish = [ "build", "twine",]

[project.scripts]
brain = "matilda_brain.cli:cli_entry"

[project.urls]
Homepage = "https://github.com/goobits/ttt"
"Bug Tracker" = "https://github.com/goobits/ttt/issues"

[tool.ruff]
line-length = 120
target-version = "py38"
cache-dir = ".artifacts/cache/ruff"

[tool.black]
line-length = 120
target-version = [ "py38",]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
cache_dir = ".artifacts/cache/mypy"

[tool.pytest.ini_options]
markers = [
    "unit: unit tests",
    "requires_api: tests that require API keys",
    "integration: integration tests", 
    "slow: slow tests",
    "benchmark: benchmark tests",
    "examples: example tests"
]
cache_dir = ".artifacts/cache/pytest"
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "-ra"
]

[tool.coverage.run]
data_file = ".artifacts/cache/.coverage"

[tool.setuptools.package-data]
matilda_brain = [ "setup.sh",]

[tool.setuptools]
zip-safe = false
include-package-data = false

[tool.setuptools.packages.find]
where = [ "src",]
include = [ "matilda_brain*",]
exclude = [ "tests*", "**/__pycache__", "*.egg-info", "build", "dist",]
